Adding MMS and preparing tests.

+++ PStoch/TP1/src/files/MM1.java
@@ -1,9 +1,8 @@
 package files;
 
 
-// TODO : if(lambda/mu < 1) -> blocage.
 public class MM1 extends queueTemplate {
-    public MM1(int lambda, int mu) throws Exception {
+    public MM1(double lambda, double mu) throws Exception {
         super(lambda, mu);
 
         if(lambda/mu < 1) throw new Exception("The parameters cause a block.");
